fix: Amplify test fails when CLI is installed from Verdaccio
The Amplify tests (used to) work like this:

- `npm create amplify`, which runs `npx create-amplify`
- Update the `package.json` to point to the correct CLI version
- `npm install` and run the Amplify CLI

However, during the first step (`npx create-amplify`) the Amplify CLI
already picks a baked-in CLI version (let's say `aws-cdk@2.1002.0`) and
`npm install`s it, before we have had a time to hack the `package.json`
file.

This works fine if `2.1002.0` is available to install, but in the case
where we have an isolated run with the candidate packages in Verdaccio,
only the release candidate is available which will have a version like
`2.1002.999`, and the first step of this test fails.

What we do instead is download the installer and change the package
version it is about to install by editing one of its data files,
`default_packages.json`.