fix: install with legacy peer deps
In the `aws-cdk-lib` pipeline, `aws-cdk-lib` will have a prerelease version:
`2.180.0-rc.0`. This doesn't satisfy the peerDependency that `aws-amplify` has
on `aws-cdk-lib` which will be a version like `^2.168.0`, some stable version.
The install fails because of that.

Try switching on `--legacy-peer-deps` to get around this (via
environment variable because Amplify runs the actual NPM command).