feat: patterns for interactive input tests (#51)
Enable integration tests to provide user interaction in order to test
interactive commands.

- Add `interact` property to `ShellOptions` that allows configuring user
input based on expected prompts.
- Detect when a subprocess requires user input by matching the expected
prompt against the last line of the process output.
- If detected, send the user input to the subprocess stdin stream.
- Add example tests

## Notes

Some code in the CLI explicitly checks if the process has a TTY, for
[example](https://github.com/aws/aws-cdk-cli/blob/0c214954e1026b85667d3052826ea902b8a486ed/packages/aws-cdk/lib/cli/cdk-toolkit.ts#L759-L761):

```typescript
if (!process.stdout.isTTY && !options.resourceMappingFile) {
  throw new ToolkitError('--resource-mapping is required when input is not a terminal');
}
```

This is reasonable, but also means this code cannot be tested unless the
CLI is started with a TTY.

To address this, we introduce
[`node-pty`](https://www.npmjs.com/package/node-pty), a package that
allows starting subprocess with a TTY attached. Now, whenever the
`interact` property is passed, the subprocess will be started with a
TTY.

Unfortunately, the return value of `pty.spawn` is not structurally
compatible (although provides similar functionality) with
`ChildProcess`, so a small abstraction is needed in order to toggle
between the two.

---

Resolves https://github.com/aws/aws-cdk/issues/31933