chore(cli-integ): remove debug integ workflow patch (#61)
As part of debugging a specific test failure in
https://github.com/aws/aws-cdk-cli-testing/pull/51, I committed a projen
patch to force the integ suite to only run a specific test. The patch
didn't work, but the PR got merged because the test was already fixed
apparently due to a prior commit I made. I just wasn't patient enough to
wait for confirmation after fixing the test.

This removes that patch.