fix: not require aws-cdk to be a default dependency of amplify (#70)
In https://github.com/aws-amplify/amplify-backend/pull/2614 Amplify is
proposing to remove the app dependency on `aws-cdk` in favor of the new
integration with `@aws-cdk/toolkit-lib`.

This would cause our tests to fail, since we expected `aws-cdk` to be
listed in the `default_packages.json` file. With this fix, we allow the
dep to be there or not.